implemented geocoding

+++ db/migrate/20190228101434_add_foreign_key_to_loans.rb
@@ -3,7 +3,7 @@ class AddForeignKeyToLoans < ActiveRecord::Migration[5.2]
     add_reference :loans, :user, foreign_key: true
     add_column :loans, :min_amount, :integer
     add_column :loans, :max_amount, :integer
-    add_column :loans, :interest_rate, :integer
+    add_column :loans, :interest_rate, :float
     add_column :loans, :payment_frequency, :string
     add_column :loans, :payment_period, :string
   end
